Re Implemented all of my original changes
