+++ src/components/dashboard/dashNav/DashboardNav.js
@@ -1,126 +1,18 @@
 import React from 'react'
-import {
-  makeStyles,
-  fade,
-  AppBar,
-  Button,
-  Toolbar,
-  Menu,
-  MenuItem,
-  InputBase
-} from '@material-ui/core'
-import IconButton from '@material-ui/core/IconButton'
-import SearchIcon from '@material-ui/icons/Search'
-import AccountCircle from '@material-ui/icons/AccountCircle'
-
-const useStyles = makeStyles(theme => ({
-  grow: {
-    flexGrow: 1
-  },
-  search: {
-    display: 'flex',
-    marginLeft: '55%', // WTF!!!
-    alignItems: 'center',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.white, 0.15),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.white, 0.25)
-    },
-    width: 'auto'
-  },
-  searchIcon: {
-    padding: theme.spacing(0, 2),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  inputRoot: {
-    color: 'inherit'
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)}px)`,
-    transition: theme.transitions.create('width'),
-    width: '100%'
-  },
-  button: {
-    marginLeft: theme.spacing(4)
-  }
-}))
+import { AppBar, Toolbar } from '@material-ui/core'
+import NavBarContent from './dashNavComponents/NavContent'
 
 // import the appBar theme from page component
 const DashNavbar = ({ style, logout }) => {
-  const classes = useStyles()
-  const [anchorEl, setAnchorEl] = React.useState(null)
-  const isMenuOpen = Boolean(anchorEl)
-
-  const handleProfileMenuOpen = event => {
-    setAnchorEl(event.currentTarget)
-  }
-
-  const handleMenuClose = () => {
-    setAnchorEl(null)
-  }
-
-  const menuId = 'primary-search-account-menu'
-  const renderMenu = (
-    <Menu
-      anchorEl={anchorEl}
-      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
-      id={menuId}
-      keepMounted
-      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
-      open={isMenuOpen}
-      onClose={handleMenuClose}
-    >
-      <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
-      <MenuItem onClick={handleMenuClose}>My account</MenuItem>
-    </Menu>
-  )
-
   return (
     <>
       <AppBar className={style}>
         <Toolbar>
           <img src='/images/Ping_Catcher.png' alt='Ping Catcher' width='9%' />
-          <div className={classes.search}>
-            <div className={classes.searchIcon}>
-              <SearchIcon />
-            </div>
-            <InputBase
-              placeholder='Search…'
-              classes={{
-                root: classes.inputRoot,
-                input: classes.inputInput
-              }}
-              inputProps={{ 'aria-label': 'search' }}
-            />
-          </div>
-          <div className={classes.grow} />
-          <div className={classes.sectionDesktop}>
-            <IconButton
-              edge='end'
-              aria-label='account of current user'
-              aria-controls={menuId}
-              aria-haspopup='true'
-              onClick={handleProfileMenuOpen}
-              color='inherit'
-            >
-              <AccountCircle />
-            </IconButton>
-            <Button className={classes.button} onClick={logout}>
-              Logout
-            </Button>
-          </div>
+          <NavBarContent />
         </Toolbar>
       </AppBar>
-      {renderMenu}
     </>
   )
 }
-
 export default DashNavbar
